2nd pass refactor. Semi tested
bugfixes
