+++ faceswap.py
@@ -1,33 +1,32 @@
 #!/usr/bin/env python3
+""" The master faceswap.py script """
 import sys
+
+import lib.cli as cli
+
 if  sys.version_info[0] < 3:
     raise Exception("This program requires at least python3.2")
 if sys.version_info[0] == 3 and sys.version_info[1] < 2:
     raise Exception("This program requires at least python3.2")
 
-from lib.cli import FullHelpArgumentParser
-
-from scripts.extract import ExtractTrainingData
-from scripts.train import TrainingProcessor
-from scripts.convert import ConvertImage
-from scripts.gui import TKGui
 
 def bad_args(args):
-    parser.print_help()
+    """ Print help on bad arguments """
+    PARSER.print_help()
     exit(0)
 
 if __name__ == "__main__":
-    parser = FullHelpArgumentParser()
-    subparser = parser.add_subparsers()
-    extract = ExtractTrainingData(
-        subparser, "extract", "Extract the faces from a pictures")
-    train = TrainingProcessor(
-        subparser, "train", "This command trains the model for the two faces A and B")
-    convert = ConvertImage(
-        subparser, "convert", "Convert a source image to a new one with the face swapped")
-    guiparsers = {'extract': extract, 'train': train, 'convert': convert}
-    gui = TKGui(
-        subparser, guiparsers, "gui", "Launch the Faceswap Graphical User Interface")
-    parser.set_defaults(func=bad_args)
-    arguments = parser.parse_args()
-    arguments.func(arguments)
+    PARSER = cli.FullHelpArgumentParser()
+    SUBPARSER = PARSER.add_subparsers()
+    EXTRACT = cli.ExtractArgs(
+        SUBPARSER, "extract", "Extract the faces from pictures")
+    TRAIN = cli.TrainArgs(
+        SUBPARSER, "train", "This command trains the model for the two faces A and B")
+    CONVERT = cli.ConvertArgs(
+        SUBPARSER, "convert", "Convert a source image to a new one with the face swapped")
+    GUIPARSERS = {'extract': EXTRACT, 'train': TRAIN, 'convert': CONVERT}
+    GUI = cli.GuiArgs(
+        SUBPARSER, "gui", "Launch the Faceswap Graphical User Interface", GUIPARSERS)
+    PARSER.set_defaults(func=bad_args)
+    ARGUMENTS = PARSER.parse_args()
+    ARGUMENTS.func(ARGUMENTS)
