+++ lib/cli.py
@@ -1,71 +1,111 @@
 #!/usr/bin python3
-""" Directory Processing Tasks """
+""" Command Line Arguments """
 
 import argparse
 import os
 import sys
-from pathlib import Path
 
-import cv2
-import numpy as np
-
-from lib.detect_blur import is_blurry
-from lib import Serializer
-from lib.utils import get_folder, get_image_paths, rotate_image, set_system_verbosity
 from plugins.PluginLoader import PluginLoader
 
-# DLIB is a GPU Memory hog, so the following modules should only be imported when required
-def import_detect_faces():
-    """ Import the faces_detect module only when it is required """
-    from lib.faces_detect import detect_faces
-    return detect_faces
+class ScriptExecutor(object):
+    """ Loads the relevant script modules and executes the script.
+        This class is initialised in each of the argparsers for the relevant command,
+        then execute script is called within their set_default function """
 
-def import_detected_face():
-    """ Import the faces_detect module only when it is required """
-    from lib.faces_detect import DetectedFace
-    return DetectedFace
+    def __init__(self, command, subparsers=None):
+        self.command = command.lower()
+        self.subparsers = subparsers
 
-def import_face_filter():
-    """ Import the FaceFilter module only when it is required """
-    from lib.FaceFilter import FaceFilter
-    return FaceFilter
+    def import_script(self):
+        """ Only import a script's modules when running that script."""
+        if self.command == 'extract':
+            from scripts.extract import Extract as script
+        elif self.command == 'train':
+            from scripts.train import Train as script
+        elif self.command == 'convert':
+            from scripts.convert import Convert as script
+        elif self.command == 'gui':
+            from scripts.gui import Gui as script
+        else:
+            script = None
+        return script
+
+    def execute_script(self, arguments):
+        """ Run the script for called command """
+        script = self.import_script()
+        args = (arguments, ) if self.command != 'gui' else (arguments, self.subparsers)
+        process = script(*args)
+        process.process()
 
 class FullPaths(argparse.Action):
     """Expand user- and relative-paths"""
-
     def __call__(self, parser, namespace, values, option_string=None):
         setattr(namespace, self.dest, os.path.abspath(
             os.path.expanduser(values)))
 
 class FullHelpArgumentParser(argparse.ArgumentParser):
-    """
-    Identical to the built-in argument parser, but on error
-    it prints full help message instead of just usage information
-    """
+    """ Identical to the built-in argument parser, but on error it
+        prints full help message instead of just usage information """
     def error(self, message):
         self.print_help(sys.stderr)
         args = {"prog": self.prog, "message": message}
         self.exit(2, "%(prog)s: error: %(message)s\n" % args)
 
-class DirectoryArgs(object):
+class FaceSwapArgs(object):
+    """ Faceswap argument parser functions that are universal
+        to all commands. Should be the parent function of all
+        subsequent argparsers """
+    def __init__(self, subparser, command, description="default", subparsers=None):
+        self.argument_list = self.get_argument_list()
+        self.optional_arguments = self.get_optional_arguments()
+        self.parser = self.create_parser(subparser, command, description)
+
+        self.add_arguments()
+
+        script = ScriptExecutor(command, subparsers)
+        self.parser.set_defaults(func=script.execute_script)
+
+    @staticmethod
+    def get_argument_list():
+        """ Put the arguments in a list so that they are accessible from both argparse and gui
+            overide for command specific arguments """
+        argument_list = []
+        return argument_list
+
+    @staticmethod
+    def get_optional_arguments():
+        """ Put the arguments in a list so that they are accessible from both argparse and gui
+            This is used for when there are sub-children (e.g. convert and extract)
+            Override this for custom arguments """
+        argument_list = []
+        return argument_list
+
+    @staticmethod
+    def create_parser(subparser, command, description):
+        """ Create the parser for the selected command """
+        parser = subparser.add_parser(
+            command,
+            help=description,
+            description=description,
+            epilog="Questions and feedback: \
+            https://github.com/deepfakes/faceswap-playground")
+        return parser
+
+    def add_arguments(self):
+        """ Parse the arguments passed in from argparse """
+        for option in self.argument_list + self.optional_arguments:
+            args = option["opts"]
+            kwargs = {key: option[key] for key in option.keys() if key != "opts"}
+            self.parser.add_argument(*args, **kwargs)
+
+
+class ExtractConvertArgs(FaceSwapArgs):
     """ This class is used as a parent class to capture arguments that
         will be used in both the extract and convert process.
 
         Arguments that can be used in both of these processes should be
         placed here, but no further processing should be done. This class
         just captures arguments """
-
-    def __init__(self, subparser, command, description="default"):
-
-        self.argument_list = self.get_argument_list()
-        self.optional_arguments = self.get_optional_arguments()
-
-        #process gets replaces in convert or extract
-        self.process = FSProcess
-        self.parser = self.create_parser(subparser, command, description)
-        self.add_arguments()
-
-        self.parser.set_defaults(func=self.execute_process)
 
     @staticmethod
     def get_argument_list():
@@ -128,389 +168,256 @@ class DirectoryArgs(object):
                               "help": "Show verbose output"})
         return argument_list
 
+class ExtractArgs(ExtractConvertArgs):
+    """ Class to parse the command line arguments for extraction.
+        Inherits base options from ExtractConvertArgs where arguments
+        that are used for both extract and convert should be placed """
+
     @staticmethod
     def get_optional_arguments():
         """ Put the arguments in a list so that they are accessible from both argparse and gui """
-        # Override this for custom arguments
         argument_list = []
+        argument_list.append({"opts": ("-r", "--rotate-images"),
+                              "type": str,
+                              "dest": "rotate_images",
+                              "default": None,
+                              "help": "If a face isn't found, rotate the images to try to "
+                                      "find a face. Can find more faces at the cost of extraction "
+                                      "speed. Pass in a single number to use increments of that "
+                                      "size up to 360, or pass in a list of numbers to enumerate "
+                                      "exactly what angles to check"})
+        argument_list.append({"opts": ("-bt", "--blur-threshold"),
+                              "type": int,
+                              "dest": "blur_thresh",
+                              "default": None,
+                              "help": "Automatically discard images blurrier than the specified "
+                                      "threshold. Discarded images are moved into a \"blurry\" "
+                                      "sub-folder. Lower values allow more blur"})
+        argument_list.append({"opts": ("-j", "--processes"),
+                              "type": int,
+                              "default": 1,
+                              "help": "Number of CPU processes to use. WARNING: ONLY USE THIS "
+                                      " IF YOU ARE NOT EXTRACTING ON A GPU. Anything above 1 "
+                                      " process on a GPU will run out of memory and will crash"})
+        argument_list.append({"opts": ("-s", "--skip-existing"),
+                              "action": "store_true",
+                              "dest": "skip_existing",
+                              "default": False,
+                              "help": "Skips frames that have already been extracted"})
+        argument_list.append({"opts": ("-dl", "--debug-landmarks"),
+                              "action": "store_true",
+                              "dest": "debug_landmarks",
+                              "default": False,
+                              "help": "Draw landmarks on the ouput faces for debug"})
+        argument_list.append({"opts": ("-ae", "--align-eyes"),
+                              "action": "store_true",
+                              "dest": "align_eyes",
+                              "default": False,
+                              "help": "Perform extra alignment to ensure left/right eyes "
+                                      "are  at the same height"})
         return argument_list
 
-    @staticmethod
-    def create_parser(subparser, command, description):
-        """ Create a directory processing parser
-            Overide for extract/convert process """
-        parser = subparser.add_parser(
-            command,
-            description=description,
-            epilog="Questions and feedback: \
-            https://github.com/deepfakes/faceswap-playground")
-        return parser
-
-    def add_arguments(self):
-        """ Parse the arguments passed in from argparse """
-        for option in self.argument_list + self.optional_arguments:
-            args = option["opts"]
-            kwargs = {key: option[key] for key in option.keys() if key != "opts"}
-            self.parser.add_argument(*args, **kwargs)
-
-    def execute_process(self, arguments):
-        """ The process to run if this command is called """
-        process = self.process(arguments)
-        process.process_wrapper()
-
-class FSProcess(object):
-    """ Class for directory processing.
-
-        This class contains functions that are used in both the extract
-        and convert process and in those processes only. If a function
-        doesn't meet that criteria then it shouldn't be placed here """
-
-    def __init__(self, arguments):
-        print('Starting, this may take a while...')
-
-        self.args = arguments
-        self.output_dir = get_folder(self.args.output_dir)
-
-        self.images = Images(self.args)
-        self.faces = Faces(self.args)
-        self.alignments = Alignments(self.args, self.faces)
-
-        self.extractor = None
-        self.export_face = True
-
-    def process_wrapper(self):
-        """ Actions to perform before and after running the extract or train process """
-        lvl = '0' if self.args.verbose else '2'
-        set_system_verbosity(lvl)
-
-        self.process()
-        self.finalize()
+class ConvertArgs(ExtractConvertArgs):
+    """ Class to parse the command line arguments for conversion.
+        Inherits base options from ExtractConvertArgs where arguments
+        that are used for both extract and convert should be placed """
 
     @staticmethod
-    def process():
-        """ Override with Extract/Convert process """
-        raise NotImplementedError
+    def get_optional_arguments():
+        """ Put the arguments in a list so that they are accessible from both argparse and gui """
+        argument_list = []
+        argument_list.append({"opts": ("-m", "--model-dir"),
+                              "action": FullPaths,
+                              "dest": "model_dir",
+                              "default": "models",
+                              "help": "Model directory. A directory containing the trained model "
+                                      "you wish to process. Defaults to 'models'"})
+        argument_list.append({"opts": ("-a", "--input-aligned-dir"),
+                              "action": FullPaths,
+                              "dest": "input_aligned_dir",
+                              "default": None,
+                              "help": "Input \"aligned directory\". A directory that should "
+                                      "contain the aligned faces extracted from the input files. "
+                                      "If you delete faces from this folder, they'll be skipped "
+                                      "during conversion. If no aligned dir is specified, all "
+                                      "faces will be converted"})
+        argument_list.append({"opts": ("-t", "--trainer"),
+                              "type": str,
+                              # case sensitive because this is used to load a plug-in.
+                              "choices": PluginLoader.get_available_models(),
+                              "default": PluginLoader.get_default_model(),
+                              "help": "Select the trainer that was used to create the model"})
+        argument_list.append({"opts": ("-c", "--converter"),
+                              "type": str,
+                              # case sensitive because this is used to load a plugin.
+                              "choices": ("Masked", "Adjust"),
+                              "default": "Masked",
+                              "help": "Converter to use"})
+        argument_list.append({"opts": ("-b", "--blur-size"),
+                              "type": int,
+                              "default": 2,
+                              "help": "Blur size. (Masked converter only)"})
+        argument_list.append({"opts": ("-e", "--erosion-kernel-size"),
+                              "dest": "erosion_kernel_size",
+                              "type": int,
+                              "default": None,
+                              "help": "Erosion kernel size. Positive values apply erosion "
+                                      "which reduces the edge of the swapped face. Negative "
+                                      "values apply dilation which allows the swapped face "
+                                      "to cover more space. (Masked converter only)"})
+        argument_list.append({"opts": ("-M", "--mask-type"),
+                              #lowercase this, because its just a string later on.
+                              "type": str.lower,
+                              "dest": "mask_type",
+                              "choices": ["rect", "facehull", "facehullandrect"],
+                              "default": "facehullandrect",
+                              "help": "Mask to use to replace faces. (Masked converter only)"})
+        argument_list.append({"opts": ("-sh", "--sharpen"),
+                              "type": str.lower,
+                              "dest": "sharpen_image",
+                              "choices": ["bsharpen", "gsharpen"],
+                              "default": None,
+                              "help": "Use Sharpen Image.bsharpen for Box Blur, gsharpen for "
+                                      "Gaussian Blur (Masked converter only)"})
+        argument_list.append({"opts": ("-g", "--gpus"),
+                              "type": int,
+                              "default": 1,
+                              "help": "Number of GPUs to use for conversion"})
+        argument_list.append({"opts": ("-fr", "--frame-ranges"),
+                              "nargs": "+",
+                              "type": str,
+                              "help": "frame ranges to apply transfer to e.g. For frames 10 to "
+                                      "50 and 90 to 100 use --frame-ranges 10-50 90-100. Files "
+                                      "must have the frame-number as the last number in the "
+                                      "name!"})
+        argument_list.append({"opts": ("-d", "--discard-frames"),
+                              "action": "store_true",
+                              "dest": "discard_frames",
+                              "default": False,
+                              "help": "When used with --frame-ranges discards frames that are "
+                                      "not processed instead of writing them out unchanged"})
+        argument_list.append({"opts": ("-s", "--swap-model"),
+                              "action": "store_true",
+                              "dest": "swap_model",
+                              "default": False,
+                              "help": "Swap the model. Instead of A -> B, swap B -> A"})
+        argument_list.append({"opts": ("-S", "--seamless"),
+                              "action": "store_true",
+                              "dest": "seamless_clone",
+                              "default": False,
+                              "help": "Use cv2's seamless clone. (Masked converter only)"})
+        argument_list.append({"opts": ("-mh", "--match-histgoram"),
+                              "action": "store_true",
+                              "dest": "match_histogram",
+                              "default": False,
+                              "help": "Use histogram matching. (Masked converter only)"})
+        argument_list.append({"opts": ("-sm", "--smooth-mask"),
+                              "action": "store_true",
+                              "dest": "smooth_mask",
+                              "default": True,
+                              "help": "Smooth mask (Adjust converter only)"})
+        argument_list.append({"opts": ("-aca", "--avg-color-adjust"),
+                              "action": "store_true",
+                              "dest": "avg_color_adjust",
+                              "default": True,
+                              "help": "Average color adjust. (Adjust converter only)"})
+        return argument_list
 
-    # EXTRACT ALIGNMENTS FROM IMAGES #
-    @staticmethod
-    def load_extractor():
-        """ Load the requested extractor for extraction """
-        # TODO Pass as argument
-        extractor_name = "Align"
-        extractor = PluginLoader.get_extractor(extractor_name)()
-
-        return extractor
-
-    def extract_face_alignments(self, filename):
-        """ Read an image from a file """
-        try:
-            image = cv2.imread(filename)
-            return filename, self.handle_image(image, filename)
-        except Exception as err:
-            if self.args.verbose:
-                print("Failed to extract from image: {}. Reason: {}".format(filename, err))
-            return filename, []
-
-    def handle_image(self, image, filename):
-        """ Attempt to extract faces from an image """
-        faces = self.faces.get_faces(image)
-        process_faces = [(idx, face) for idx, face in faces]
-        # Run image rotator if requested and no faces found
-        if self.images.rotation_angles and not process_faces:
-            process_faces, image = self.rotate_frame(image)
-
-        return [self.process_single_face(idx, face, filename, image)
-                for idx, face in process_faces]
-
-    def process_single_face(self, idx, face, filename, image):
-        """ Perform processing on found faces """
-        output_file = self.output_dir / Path(filename).stem if self.export_face else None
-
-        # Draws landmarks for debug
-        if hasattr(self.args, 'debug_landmarks') and self.args.debug_landmarks:
-            self.faces.draw_landmarks_on_face(face, image)
-
-        align_eyes = self.args.align_eyes if hasattr(self.args, 'align_eyes') else False
-        resized_image, t_mat = self.extractor.extract(image, face, 256, align_eyes)
-
-        # Detect blurry images
-        if hasattr(self.args, 'blur_thresh') and self.args.blur_thresh:
-            blurry_file = self.detect_blurry_faces(face, t_mat, resized_image, filename)
-            output_file = blurry_file if blurry_file else output_file
-
-        if self.export_face:
-            cv2.imwrite("{}_{}{}".format(str(output_file),
-                                         str(idx),
-                                         Path(filename).suffix),
-                        resized_image)
-
-        return {"r": face.r,
-                "x": face.x,
-                "w": face.w,
-                "y": face.y,
-                "h": face.h,
-                "landmarksXY": face.landmarksAsXY()}
-
-    def rotate_frame(self, image):
-        """ Rotate the image to extract more faces if requested """
-        for angle in self.images.rotation_angles:
-            rotated_image = rotate_image(image, angle)
-            faces = self.faces.get_faces(rotated_image, rotation=angle)
-            rotated_faces = [(idx, face) for idx, face in faces]
-            if rotated_faces and self.args.verbose:
-                print("found face(s) by rotating image {} degrees".format(angle))
-            if rotated_faces:
-                break
-            return rotated_faces, rotated_image
-
-    def detect_blurry_faces(self, face, t_mat, resized_image, filename):
-        """ Detect and move blurry face """
-        blurry_file = None
-        aligned_landmarks = self.extractor.transform_points(face.landmarksAsXY(), t_mat, 256, 48)
-        feature_mask = self.extractor.get_feature_mask(aligned_landmarks / 256, 256, 48)
-        feature_mask = cv2.blur(feature_mask, (10, 10))
-        isolated_face = cv2.multiply(feature_mask, resized_image.astype(float)).astype(np.uint8)
-        blurry, focus_measure = is_blurry(isolated_face, self.args.blur_thresh)
-
-        if blurry:
-            print("{}'s focus measure of {} was below the blur threshold, "
-                  "moving to \"blurry\"".format(Path(filename).stem, focus_measure))
-            blurry_file = get_folder(Path(self.output_dir) / Path("blurry")) / \
-                Path(filename).stem
-        return blurry_file
-
-    def finalize(self):
-        """ Finalize the image processing """
-        print("-------------------------")
-        print("Images found:        {}".format(self.images.images_found))
-        print("Faces detected:      {}".format(self.faces.num_faces_detected))
-        print("-------------------------")
-
-        if self.faces.verify_output:
-            print("Note:")
-            print("Multiple faces were detected in one or more pictures.")
-            print("Double check your results.")
-            print("-------------------------")
-
-        self.images.images_found = 0
-        self.faces.num_faces_detected = 0
-        print("Done!")
-
-class Images(object):
-    """ Holds the full frames/images """
-    def __init__(self, arguments):
-        self.args = arguments
-        self.rotation_angles = self.get_rotation_angles()
-        self.already_processed = self.get_already_processed()
-        self.input_images = self.get_input_images()
-
-        self.images_found = 0
-
-    def get_rotation_angles(self):
-        """ Set the rotation angles. Includes backwards compatibility for the 'on'
-            and 'off' options:
-                - 'on' - increment 90 degrees
-                - 'off' - disable """
-        if (not hasattr(self.args, 'rotate_images')
-                or not self.args.rotate_images
-                or self.args.rotate_images == "off"):
-            rotation_angles = None
-        elif self.rotation_angles == "on":
-            rotation_angles = range(90, 360, 90)
-        else:
-            rotation_angles = [int(angle)
-                               for angle in self.args.rotate_images.split(",")]
-            if len(rotation_angles) == 1:
-                rotation_step_size = rotation_angles[0]
-                rotation_angles = range(rotation_step_size, 360, rotation_step_size)
-            elif len(rotation_angles) > 1:
-                rotation_angles = rotation_angles
-            else:
-                rotation_angles = None
-        return rotation_angles
-
-    def get_already_processed(self):
-        """ Return the images that already exist in the output directory """
-        print("Output Directory: {}".format(self.args.output_dir))
-
-        if not hasattr(self.args, 'skip_existing') or not self.args.skip_existing:
-            return None
-
-        return get_image_paths(self.args.output_dir)
-
-    def get_input_images(self):
-        """ Return the list of images that are to be processed """
-        if not os.path.exists(self.args.input_dir):
-            print("Input directory {} not found.".format(self.args.input_dir))
-            exit(1)
-
-        print("Input Directory: {}".format(self.args.input_dir))
-
-        if hasattr(self.args, 'skip_existing') and self.args.skip_existing:
-            input_images = get_image_paths(self.args.input_dir, self.already_processed)
-            print("Excluding %s files" % len(self.already_processed))
-        else:
-            input_images = get_image_paths(self.args.input_dir)
-
-        return input_images
-
-    def read_directory(self):
-        """ Return number of images from directory for tqdm """
-        self.images_found = len(self.input_images)
-        return self.input_images
-
-class Faces(object):
-    """ Holds the faces """
-    def __init__(self, arguments):
-        self.args = arguments
-        self.filter = self.load_face_filter()
-
-        self.faces_detected = dict()
-        self.num_faces_detected = 0
-        self.verify_output = False
-
-    def have_face(self, filename):
-        """ return path of images that have faces """
-        return os.path.basename(filename) in self.faces_detected
-
-    def load_face_filter(self):
-        """ Load faces to filter out of images """
-        facefilter = None
-        filter_files = [self.set_face_filter(filter_type)
-                        for filter_type in ('filter', 'nfilter')]
-
-        if any(filters for filters in filter_files):
-            FaceFilter = import_face_filter()
-            facefilter = FaceFilter(filter_files[0], filter_files[1], self.args.ref_threshold)
-        return facefilter
-
-    def set_face_filter(self, filter_list):
-        """ Set the required filters """
-        filter_files = list()
-        filter_args = getattr(self.args, filter_list)
-        if filter_args:
-            print("{}: {}".format(filter_list.title(), filter_args))
-            filter_files = filter_args
-            if not isinstance(filter_args, list):
-                filter_files = [filter_args]
-            filter_files = list(filter(lambda fnc: Path(fnc).exists(), filter_files))
-        return filter_files
-
-    def get_faces(self, image, rotation=0):
-        """ Extract the faces from an image """
-        detect_faces = import_detect_faces()
-        faces_count = 0
-        faces = detect_faces(image, self.args.detector, self.args.verbose, rotation)
-
-        for face in faces:
-            if self.filter and not self.filter.check(face):
-                if self.args.verbose:
-                    print("Skipping not recognized face!")
-                continue
-            yield faces_count, face
-
-            self.num_faces_detected += 1
-            faces_count += 1
-
-        if faces_count > 1 and self.args.verbose:
-            self.verify_output = True
-
-    def get_faces_alignments(self, filename, image):
-        """ Retrieve the face alignments from an image """
-        DetectedFace = import_detected_face()
-        faces_count = 0
-        faces = self.faces_detected[os.path.basename(filename)]
-        for rawface in faces:
-            face = DetectedFace(**rawface)
-            # Rotate the image if necessary
-            if face.r != 0:
-                image = rotate_image(image, face.r)
-            face.image = image[face.y : face.y + face.h, face.x : face.x + face.w]
-            if self.filter and not self.filter.check(face):
-                if self.args.verbose:
-                    print("Skipping not recognized face!")
-                continue
-
-            yield faces_count, face
-            self.num_faces_detected += 1
-            faces_count += 1
-        if faces_count > 1 and self.args.verbose:
-            print("Note: Found more than one face in an image! File: %s" % filename)
-            self.verify_output = True
+class TrainArgs(FaceSwapArgs):
+    """ Class to parse the command line arguments for training """
 
     @staticmethod
-    def draw_landmarks_on_face(face, image):
-        """ Draw debug landmarks on extracted face """
-        for (pos_x, pos_y) in face.landmarksAsXY():
-            cv2.circle(image, (pos_x, pos_y), 2, (0, 0, 255), -1)
+    def get_argument_list():
+        """ Put the arguments in a list so that they are accessible from both argparse and gui """
+        argument_list = []
+        argument_list.append({"opts": ("-A", "--input-A"),
+                              "action": FullPaths,
+                              "dest": "input_A",
+                              "default": "input_A",
+                              "help": "Input directory. A directory containing training images "
+                                      "for face A. Defaults to 'input'"})
+        argument_list.append({"opts": ("-B", "--input-B"),
+                              "action": FullPaths,
+                              "dest": "input_B",
+                              "default": "input_B",
+                              "help": "Input directory. A directory containing training images "
+                                      "for face B Defaults to 'input'"})
+        argument_list.append({"opts": ("-m", "--model-dir"),
+                              "action": FullPaths,
+                              "dest": "model_dir",
+                              "default": "models",
+                              "help": "Model directory. This is where the training data will "
+                                      "be stored. Defaults to 'model'"})
+        argument_list.append({"opts": ("-s", "--save-interval"),
+                              "type": int,
+                              "dest": "save_interval",
+                              "default": 100,
+                              "help": "Sets the number of iterations before saving the model"})
+        argument_list.append({"opts": ("-t", "--trainer"),
+                              "type": str,
+                              "choices": PluginLoader.get_available_models(),
+                              "default": PluginLoader.get_default_model(),
+                              "help": "Select which trainer to use, Use LowMem for cards with "
+                                      " less than 2GB of VRAM"})
+        argument_list.append({"opts": ("-bs", "--batch-size"),
+                              "type": int,
+                              "default": 64,
+                              "help": "Batch size, as a power of 2 (64, 128, 256, etc)"})
+        argument_list.append({"opts": ("-ep", "--epochs"),
+                              "type": int,
+                              "default": 1000000,
+                              "help": "Length of training in epochs"})
+        argument_list.append({"opts": ("-g", "--gpus"),
+                              "type": int,
+                              "default": 1,
+                              "help": "Number of GPUs to use for training"})
+        argument_list.append({"opts": ("-p", "--preview"),
+                              "action": "store_true",
+                              "dest": "preview",
+                              "default": False,
+                              "help": "Show preview output. If not specified, write progress "
+                                      "to file"})
+        argument_list.append({"opts": ("-w", "--write-image"),
+                              "action": "store_true",
+                              "dest": "write_image",
+                              "default": False,
+                              "help": "Writes the training result to a file even on "
+                                      "preview mode"})
+        argument_list.append({"opts": ("-pl", "--use-perceptual-loss"),
+                              "action": "store_true",
+                              "dest": "perceptual_loss",
+                              "default": False,
+                              "help": "Use perceptual loss while training"})
+        argument_list.append({"opts": ("-ag", "--allow-growth"),
+                              "action": "store_true",
+                              "dest": "allow_growth",
+                              "default": False,
+                              "help": "Sets allow_growth option of Tensorflow to spare memory "
+                                      "on some configs"})
+        argument_list.append({"opts": ("-v", "--verbose"),
+                              "action": "store_true",
+                              "dest": "verbose",
+                              "default": False,
+                              "help": "Show verbose output"})
+        # This is a hidden argument to indicate that the GUI is being used,
+        # so the preview window should be redirected Accordingly
+        argument_list.append({"opts": ("-gui", "--gui"),
+                              "action": "store_true",
+                              "dest": "redirect_gui",
+                              "default": False,
+                              "help": argparse.SUPPRESS})
+        return argument_list
 
-class Alignments(object):
-    """ Holds processes pertaining to the alignments file """
-    def __init__(self, arguments, faces):
-        self.args = arguments
-        self.faces = faces
+class GuiArgs(FaceSwapArgs):
+    """ Class to parse the command line arguments for training """
 
-        self.serializer = self.get_serializer()
-
-    def get_serializer(self):
-        """ Set the serializer to be used for loading and saving alignments """
-        if not self.args.serializer and self.args.alignments_path:
-            ext = os.path.splitext(self.args.alignments_path)[-1]
-            serializer = Serializer.get_serializer_fromext(ext)
-            print("Alignments Output: {}".format(self.args.alignments_path))
-        else:
-            serializer = Serializer.get_serializer(self.args.serializer or "json")
-        print("Using {} serializer".format(serializer.ext))
-        return serializer
-
-    def get_alignments_path(self):
-        """ Return the path to alignments file """
-        if self.args.alignments_path:
-            alignfile = self.args.alignments_path
-        else:
-            alignfile = os.path.join(str(self.args.input_dir),
-                                     "alignments.{}".format(self.serializer.ext))
-        print("Alignments filepath: %s" % alignfile)
-        return alignfile
-
-    def read_alignments(self):
-        """ Read the serialized alignments file """
-        alignfile = self.get_alignments_path()
-        try:
-            with open(alignfile, self.serializer.roptions) as align:
-                self.faces.faces_detected = self.serializer.unmarshal(align.read())
-        except Exception as err:
-            print("{} not read!".format(alignfile))
-            print(str(err))
-            self.faces.faces_detected = dict()
-
-    def write_alignments(self):
-        """ Write the serialized alignments file """
-        alignfile = self.get_alignments_path()
-
-        if hasattr(self.args, 'skip_existing') and self.args.skip_existing:
-            self.load_skip_alignments(alignfile)
-
-        try:
-            print("Writing alignments to: {}".format(alignfile))
-            with open(alignfile, self.serializer.woptions) as align:
-                align.write(self.serializer.marshal(self.faces.faces_detected))
-        except Exception as err:
-            print("{} not written!".format(alignfile))
-            print(str(err))
-            self.faces.faces_detected = dict()
-
-    def load_skip_alignments(self, alignfile):
-        """ Load existing alignments if skipping existing images """
-        if self.have_alignments():
-            with open(alignfile, self.serializer.roptions) as inf:
-                data = self.serializer.unmarshal(inf.read())
-                for key, val in data.items():
-                    self.faces.faces_detected[key] = val
-        else:
-            print("Existing alignments file '%s' not found." % alignfile)
-
-    def have_alignments(self):
-        """ Check if an alignments file exists """
-        alignfile = self.get_alignments_path()
-        return os.path.exists(alignfile)
+    @staticmethod
+    def get_argument_list():
+        """ Put the arguments in a list so that they are accessible from both argparse and gui """
+        argument_list = []
+        argument_list.append({"opts": ("-d", "--debug"),
+                              "action": "store_true",
+                              "dest": "debug",
+                              "default": False,
+                              "help": "Output to Shell console instead of GUI console"})
+        return argument_list
