+++ scripts/extract.py
@@ -2,109 +2,118 @@
 """ The script to run the extract process of faceswap """
 
 import os
+from pathlib import Path
 
 from tqdm import tqdm
 
-from lib.cli import DirectoryArgs, FSProcess
 from lib.multithreading import pool_process
+from lib.utils import set_system_verbosity
+from scripts.inout import Alignments, Faces, Images, Utils
 
-class ExtractTrainingData(DirectoryArgs):
-    """ Class to parse the command line arguments for extraction.
-        Inherits base options from lib.DirectoryArgs """
-
-    @staticmethod
-    def get_optional_arguments():
-        """ Put the arguments in a list so that they are accessible from both argparse and gui """
-        argument_list = []
-        argument_list.append({"opts": ("-r", "--rotate-images"),
-                              "type": str,
-                              "dest": "rotate_images",
-                              "default": None,
-                              "help": "If a face isn't found, rotate the images to try to "
-                                      "find a face. Can find more faces at the cost of extraction "
-                                      "speed. Pass in a single number to use increments of that "
-                                      "size up to 360, or pass in a list of numbers to enumerate "
-                                      "exactly what angles to check"})
-        argument_list.append({"opts": ("-bt", "--blur-threshold"),
-                              "type": int,
-                              "dest": "blur_thresh",
-                              "default": None,
-                              "help": "Automatically discard images blurrier than the specified "
-                                      "threshold. Discarded images are moved into a \"blurry\" "
-                                      "sub-folder. Lower values allow more blur"})
-        argument_list.append({"opts": ("-j", "--processes"),
-                              "type": int,
-                              "default": 1,
-                              "help": "Number of CPU processes to use. WARNING: ONLY USE THIS "
-                                      " IF YOU ARE NOT EXTRACTING ON A GPU. Anything above 1 "
-                                      " process on a GPU will run out of memory and will crash"})
-        argument_list.append({"opts": ("-s", "--skip-existing"),
-                              "action": "store_true",
-                              "dest": "skip_existing",
-                              "default": False,
-                              "help": "Skips frames that have already been extracted"})
-        argument_list.append({"opts": ("-dl", "--debug-landmarks"),
-                              "action": "store_true",
-                              "dest": "debug_landmarks",
-                              "default": False,
-                              "help": "Draw landmarks on the ouput faces for debug"})
-        argument_list.append({"opts": ("-ae", "--align-eyes"),
-                              "action": "store_true",
-                              "dest": "align_eyes",
-                              "default": False,
-                              "help": "Perform extra alignment to ensure left/right eyes "
-                                      "are  at the same height"})
-        return argument_list
-
-    def create_parser(self, subparser, command, description):
-        """ Create the extract parser """
-        self.optional_arguments = self.get_optional_arguments()
-        self.process = Extract
-
-        parser = subparser.add_parser(
-            command,
-            help="Extract the faces from a pictures.",
-            description=description,
-            epilog="Questions and feedback: \
-            https://github.com/deepfakes/faceswap-playground")
-        return parser
-
-class Extract(FSProcess):
-    """ The extract process. Inherits from cli.FSProcess, including the additional
-        classes: images, faces, alignments
-
-        As most of the extract processes are required for Convert in the case when
-        there are no alignments available, most of the processes will be found in
-        lib.cli.py """
+class Extract(object):
+    """ The extract process. """
 
     def __init__(self, arguments):
-        FSProcess.__init__(self, arguments)
-        self.extractor = self.load_extractor()
+        self.args = arguments
+
+        self.images = Images(self.args)
+        self.faces = Faces(self.args)
+        self.alignments = Alignments(self.args)
+
+        self.output_dir = self.faces.output_dir
+
         self.export_face = True
+
+    def set_defaults(self):
+        """ Set default arguments.
+
+        The Extract process may be called by convert (if alignments are
+        not found). Not all options are available, so the arguments need
+        to be set to default values for those that are not available """
+        pass
 
     def process(self):
         """ Perform the extraction process """
-        processes = self.args.processes
+        print('Starting, this may take a while...')
+        Utils.set_verbosity(self.args.verbose)
 
-        if processes != 1:
-            self.multi_process(processes)
+        if hasattr(self.args, 'processes') and self.args.processes > 1:
+            self.extract_multi_process()
         else:
-            self.single_process()
+            self.extract_single_process()
 
-        self.alignments.write_alignments()
+        self.alignments.write_alignments(self.faces.faces_detected)
 
-    def multi_process(self, processes):
-        """ Run extraction in a multiple processes """
-        files = list(self.images.read_directory())
-        for filename, faces in tqdm(pool_process(self.extract_face_alignments,
-                                                 files,
-                                                 processes=processes),
-                                    total=len(files)):
+        images, faces = Utils.finalize(self.images.images_found,
+                                       self.faces.num_faces_detected,
+                                       self.faces.verify_output)
+        self.images.images_found = images
+        self.faces.num_faces_detected = faces
+
+    def extract_single_process(self):
+        """ Run extraction in a single process """
+        for filename in tqdm(self.images.input_images):
+            filename, faces = self.process_single_image(filename)
+            self.faces.faces_detected[os.path.basename(filename)] = faces
+
+    def extract_multi_process(self):
+        """ Run the extraction on the correct number of processes """
+        for filename, faces in tqdm(pool_process(self.process_single_image,
+                                                 self.images.input_images,
+                                                 processes=self.args.processes),
+                                    total=self.images.images_found):
             self.faces.num_faces_detected += 1
             self.faces.faces_detected[os.path.basename(filename)] = faces
 
-    def single_process(self):
-        """ Run extraction in a single process """
-        for filename in tqdm(self.images.read_directory()):
-            filename, faces = self.extract_face_alignments(filename)
-            self.faces.faces_detected[os.path.basename(filename)] = faces
+    def process_single_image(self, filename):
+        """ Detect faces in an image. Rotate the image the specified amount
+            until at least one face is found, or until image rotations are
+            depleted.
+            Once at least one face has been detected, pass to process_single_face
+            to process the individual faces """
+        retval = filename, list()
+        try:
+            image = self.images.cv2_read_write('read', filename)
+
+            for angle in self.images.rotation_angles:
+                image = Utils.rotate_image_by_angle(image, angle)
+                faces = self.faces.get_faces(image, angle)
+                process_faces = [(idx, face) for idx, face in faces]
+                if process_faces and angle != 0 and self.args.verbose:
+                    print("found face(s) by rotating image {} degrees".format(angle))
+                if process_faces:
+                    break
+
+            final_faces = [self.process_single_face(idx, face, filename, image)
+                           for idx, face in process_faces]
+
+            retval = filename, final_faces
+        except Exception as err:
+            if self.args.verbose:
+                print("Failed to extract from image: {}. Reason: {}".format(filename, err))
+        return retval
+
+    def process_single_face(self, idx, face, filename, image):
+        """ Perform processing on found faces """
+        output_file = self.output_dir / Path(filename).stem if self.export_face else None
+
+        self.faces.draw_landmarks_on_face(face, image)
+
+        resized_face, t_mat = self.faces.extractor.extract(image,
+                                                           face,
+                                                           256,
+                                                           self.faces.align_eyes)
+
+        blurry_file = self.faces.detect_blurry_faces(face, t_mat, resized_face, filename)
+        output_file = blurry_file if blurry_file else output_file
+
+        if self.export_face:
+            filename = "{}_{}{}".format(str(output_file), str(idx), Path(filename).suffix)
+            self.images.cv2_read_write('write', filename, resized_face)
+
+        return {"r": face.r,
+                "x": face.x,
+                "w": face.w,
+                "y": face.y,
+                "h": face.h,
+                "landmarksXY": face.landmarksAsXY()}
