+++ scripts/train.py
@@ -1,144 +1,17 @@
 #!/usr/bin python3
 """ The script to run the training process of faceswap """
 
+import os
+import sys
 import threading
 
 import cv2
+import tensorflow as tf
+from keras.backend.tensorflow_backend import set_session
 
-from lib.cli import FullPaths, argparse, os, sys
 from lib.utils import get_folder, get_image_paths, set_system_verbosity
 from plugins.PluginLoader import PluginLoader
 
-def import_tensorflow_keras():
-    """ Import the TensorFlow and keras set_session modules only when they are required """
-    import tensorflow as tflow
-    from keras.backend.tensorflow_backend import set_session
-    return (tflow, set_session)
-
-class TrainingProcessor(object):
-    """ Class to parse the command line arguments for training """
-
-    def __init__(self, subparser, command, description="default"):
-        self.argument_list = self.get_argument_list()
-        self.optional_arguments = self.get_optional_arguments()
-
-        self.parser = self.create_parser(subparser, command, description)
-
-        self.add_arguments()
-        self.parser.set_defaults(func=self.execute_process)
-
-    @staticmethod
-    def get_argument_list():
-        """ Put the arguments in a list so that they are accessible from both argparse and gui """
-        argument_list = []
-        argument_list.append({"opts": ("-A", "--input-A"),
-                              "action": FullPaths,
-                              "dest": "input_A",
-                              "default": "input_A",
-                              "help": "Input directory. A directory containing training images "
-                                      "for face A. Defaults to 'input'"})
-        argument_list.append({"opts": ("-B", "--input-B"),
-                              "action": FullPaths,
-                              "dest": "input_B",
-                              "default": "input_B",
-                              "help": "Input directory. A directory containing training images "
-                                      "for face B Defaults to 'input'"})
-        argument_list.append({"opts": ("-m", "--model-dir"),
-                              "action": FullPaths,
-                              "dest": "model_dir",
-                              "default": "models",
-                              "help": "Model directory. This is where the training data will "
-                                      "be stored. Defaults to 'model'"})
-        argument_list.append({"opts": ("-s", "--save-interval"),
-                              "type": int,
-                              "dest": "save_interval",
-                              "default": 100,
-                              "help": "Sets the number of iterations before saving the model"})
-        argument_list.append({"opts": ("-t", "--trainer"),
-                              "type": str,
-                              "choices": PluginLoader.get_available_models(),
-                              "default": PluginLoader.get_default_model(),
-                              "help": "Select which trainer to use, Use LowMem for cards with "
-                                      " less than 2GB of VRAM"})
-        argument_list.append({"opts": ("-bs", "--batch-size"),
-                              "type": int,
-                              "default": 64,
-                              "help": "Batch size, as a power of 2 (64, 128, 256, etc)"})
-        argument_list.append({"opts": ("-ep", "--epochs"),
-                              "type": int,
-                              "default": 1000000,
-                              "help": "Length of training in epochs"})
-        argument_list.append({"opts": ("-g", "--gpus"),
-                              "type": int,
-                              "default": 1,
-                              "help": "Number of GPUs to use for training"})
-        argument_list.append({"opts": ("-p", "--preview"),
-                              "action": "store_true",
-                              "dest": "preview",
-                              "default": False,
-                              "help": "Show preview output. If not specified, write progress "
-                                      "to file"})
-        argument_list.append({"opts": ("-w", "--write-image"),
-                              "action": "store_true",
-                              "dest": "write_image",
-                              "default": False,
-                              "help": "Writes the training result to a file even on "
-                                      "preview mode"})
-        argument_list.append({"opts": ("-pl", "--use-perceptual-loss"),
-                              "action": "store_true",
-                              "dest": "perceptual_loss",
-                              "default": False,
-                              "help": "Use perceptual loss while training"})
-        argument_list.append({"opts": ("-ag", "--allow-growth"),
-                              "action": "store_true",
-                              "dest": "allow_growth",
-                              "default": False,
-                              "help": "Sets allow_growth option of Tensorflow to spare memory "
-                                      "on some configs"})
-        argument_list.append({"opts": ("-v", "--verbose"),
-                              "action": "store_true",
-                              "dest": "verbose",
-                              "default": False,
-                              "help": "Show verbose output"})
-        # This is a hidden argument to indicate that the GUI is being used,
-        # so the preview window should be redirected Accordingly
-        argument_list.append({"opts": ("-gui", "--gui"),
-                              "action": "store_true",
-                              "dest": "redirect_gui",
-                              "default": False,
-                              "help": argparse.SUPPRESS})
-        return argument_list
-
-    @staticmethod
-    def create_parser(subparser, command, description):
-        """ Create the train parser """
-        parser = subparser.add_parser(
-            command,
-            help="This command trains the model for the two faces A and B.",
-            description=description,
-            epilog="Questions and feedback: \
-            https://github.com/deepfakes/faceswap-playground")
-        return parser
-
-    @staticmethod
-    def get_optional_arguments():
-        """ Put the arguments in a list so that they are accessible from both argparse and gui """
-        # Override this for custom arguments
-        argument_list = []
-        return argument_list
-
-    def add_arguments(self):
-        """ Parse the arguments passed in from argparse """
-        for option in self.argument_list + self.optional_arguments:
-            args = option["opts"]
-            kwargs = {key: option[key] for key in option.keys() if key != "opts"}
-            self.parser.add_argument(*args, **kwargs)
-
-    @staticmethod
-    def execute_process(arguments):
-        """ The process to run if this command is called """
-        process = Train(arguments)
-        process.process()
 
 class Train(object):
     """ The training process.  """
@@ -299,11 +172,10 @@ class Train(object):
     @staticmethod
     def set_tf_allow_growth():
         """ Allow TensorFlow to manage VRAM growth """
-        tflow, set_session = import_tensorflow_keras()
-        config = tflow.ConfigProto()
+        config = tf.ConfigProto()
         config.gpu_options.allow_growth = True
         config.gpu_options.visible_device_list = "0"
-        set_session(tflow.Session(config=config))
+        set_session(tf.Session(config=config))
 
     def show(self, image, name=""):
         """ Generate the preview and write preview file output """
